[FIX] Fixing warning in `bam_stats.cpp` in MSVC++.
git-svn-id: http://svn.seqan.de/seqan/trunk@11967 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/extras/demos/bam_stats.cpp
@@ -277,7 +277,7 @@ int realignBamRecord(TReference & reference, BamAlignmentRecord & record, unsign
     TPatternState patternState;
     
     int distance = maxIndels + 1;
-    while (find(finder, readSeq, patternState, -maxIndels))
+    while (find(finder, readSeq, patternState, -static_cast<int>(maxIndels)))
         distance = std::min(distance, -getScore(patternState));
     
     return distance;
